requirements.txt
Django==4.1.7
django-environ==0.10.0
psycopg2==2.9.5
django-rest-framework==0.1.0
Pillow==9.4.0
